Outsourced the media slider
