+++ gamevault/UserControls/GameViewUserControl.xaml
@@ -5,8 +5,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:conv="clr-namespace:gamevault.Converter"
              xmlns:helper="clr-namespace:gamevault.Helper"
-             xmlns:mah="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"
-             xmlns:wv2="clr-namespace:Microsoft.Web.WebView2.Wpf;assembly=Microsoft.Web.WebView2.Wpf"
+             xmlns:mah="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"             
              xmlns:local="clr-namespace:gamevault.UserControls"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800" Loaded="UserControl_Loaded" KeyDown="ReloadGameView_Click" Focusable="True" FocusVisualStyle="{x:Null}">
@@ -422,72 +421,7 @@
                         </Grid>
                     </StackPanel>
                     <!--MEDIA SLIDER-->
-                    <Grid x:Name="uiWebViewControlHost" Panel.ZIndex="1" Background="Transparent" Height="500" Width="980" HorizontalAlignment="Center" VerticalAlignment="Bottom" Tag="NoPopupTransition" KeyDown="MediaSliderFullscreen_Escape_KeyDown">
-                        <Border HorizontalAlignment="Right" VerticalAlignment="Top" Margin="0,0,-15,0" Width="60">
-                            <TextBlock x:Name="uiTxtMediaIndex" FontSize="16" Foreground="White" HorizontalAlignment="Left"/>
-                        </Border>
-                        <Grid Background="Black" Margin="46,0,46,0">
-                            <Button Style="{StaticResource ButtonWrapper}" Cursor="Hand" VerticalAlignment="Center" HorizontalAlignment="Center" Width="35" Height="35" Margin="0,0,3,0" Click="ReloadMediaSlider_Click">
-                                <Grid Style="{DynamicResource HoverEffect}" RenderTransformOrigin="0.5,0.5" Background="#02000000">
-                                    <Path Data="{StaticResource IconReload}" Fill="#7AFFFFFF">
-                                        <Path.RenderTransform>
-                                            <ScaleTransform ScaleX="1.5" ScaleY="1.5"/>
-                                        </Path.RenderTransform>
-                                    </Path>
-                                    <Grid.RenderTransform>
-                                        <ScaleTransform/>
-                                    </Grid.RenderTransform>
-                                </Grid>
-                            </Button>
-                            <wv2:WebView2 x:Name="uiWebView" />
-                        </Grid>
-                        <Button Style="{StaticResource ButtonWrapper}" VerticalAlignment="Center" HorizontalAlignment="Right" Cursor="Hand" Width="35" Height="35" Margin="0,0,3,40" Click="PrevMedia_Click">
-                            <Grid Style="{DynamicResource HoverEffect}" Background="#02000000">
-                                <Path Data="{StaticResource IconArrowBordered}" Fill="White" RenderTransformOrigin="0.45,0.32">
-                                    <Path.RenderTransform>
-                                        <TransformGroup>
-                                            <ScaleTransform ScaleX="1.5" ScaleY="1.5"/>
-                                            <SkewTransform/>
-                                            <RotateTransform Angle="-90"/>
-                                            <TranslateTransform/>
-                                        </TransformGroup>
-                                    </Path.RenderTransform>
-                                </Path>
-                                <Grid.RenderTransform>
-                                    <ScaleTransform/>
-                                </Grid.RenderTransform>
-                            </Grid>
-                        </Button>
-                        <Button Style="{StaticResource ButtonWrapper}" Cursor="Hand" VerticalAlignment="Center" HorizontalAlignment="Right" Width="35" Height="35" Margin="0,40,3,0" Click="NextMedia_Click">
-                            <Grid Style="{DynamicResource HoverEffect}" Background="#02000000">
-                                <Path Data="{StaticResource IconArrowBordered}" Fill="White" RenderTransformOrigin="0.32,0.45">
-                                    <Path.RenderTransform>
-                                        <TransformGroup>
-                                            <ScaleTransform ScaleX="1.5" ScaleY="1.5"/>
-                                            <SkewTransform/>
-                                            <RotateTransform Angle="90"/>
-                                            <TranslateTransform/>
-                                        </TransformGroup>
-                                    </Path.RenderTransform>
-                                </Path>
-                                <Grid.RenderTransform>
-                                    <ScaleTransform/>
-                                </Grid.RenderTransform>
-                            </Grid>
-                        </Button>
-                        <Button Style="{StaticResource ButtonWrapper}" Cursor="Hand" VerticalAlignment="Bottom" HorizontalAlignment="Right" Width="35" Height="35" Margin="0,0,3,0" Click="MediaSliderFullscreen_Click">
-                            <Grid Style="{DynamicResource HoverEffect}" RenderTransformOrigin="0.5,0.5" Background="#02000000">
-                                <Path Data="{StaticResource IconGameViewMaximize}" Fill="White" RenderTransformOrigin="-0.05,-0.05">
-                                    <Path.RenderTransform>
-                                        <ScaleTransform ScaleX="2" ScaleY="2"/>
-                                    </Path.RenderTransform>
-                                </Path>
-                                <Grid.RenderTransform>
-                                    <ScaleTransform/>
-                                </Grid.RenderTransform>
-                            </Grid>
-                        </Button>
-                    </Grid>
+                    <local:MediaSlider x:Name="uiMediaSlider" Panel.ZIndex="1" Background="Transparent" Height="500" Width="980" HorizontalAlignment="Center" VerticalAlignment="Bottom" Tag="NoPopupTransition"/>
                     <!--PROGRESS-->
                     <StackPanel HorizontalAlignment="Right" Margin="0,0,0,0">
                         <StackPanel Orientation="Horizontal" Margin="0,-5,0,0">
