+++ gamevault/UserControls/GeneralControls/MediaSlider.xaml
@@ -1,0 +1,76 @@
+﻿<UserControl x:Class="gamevault.UserControls.MediaSlider"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:wv2="clr-namespace:Microsoft.Web.WebView2.Wpf;assembly=Microsoft.Web.WebView2.Wpf"
+             xmlns:local="clr-namespace:gamevault.UserControls"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800" Focusable="True">
+    <Grid KeyDown="MediaSliderFullscreen_Escape_KeyDown">
+        <Border HorizontalAlignment="Right" VerticalAlignment="Top" Margin="0,0,-15,0" Width="60">
+            <TextBlock x:Name="uiTxtMediaIndex" FontSize="16" Foreground="White" HorizontalAlignment="Left"/>
+        </Border>
+        <Grid Background="Black" Margin="46,0,46,0">
+            <Button Style="{StaticResource ButtonWrapper}" Cursor="Hand" VerticalAlignment="Center" HorizontalAlignment="Center" Width="35" Height="35" Margin="0,0,3,0" Click="ReloadMediaSlider_Click">
+                <Grid Style="{DynamicResource HoverEffect}" RenderTransformOrigin="0.5,0.5" Background="#02000000">
+                    <Path Data="{StaticResource IconReload}" Fill="#7AFFFFFF">
+                        <Path.RenderTransform>
+                            <ScaleTransform ScaleX="1.5" ScaleY="1.5"/>
+                        </Path.RenderTransform>
+                    </Path>
+                    <Grid.RenderTransform>
+                        <ScaleTransform/>
+                    </Grid.RenderTransform>
+                </Grid>
+            </Button>
+            <wv2:WebView2 x:Name="uiWebView" />
+        </Grid>
+        <Button Style="{StaticResource ButtonWrapper}" VerticalAlignment="Center" HorizontalAlignment="Right" Cursor="Hand" Width="35" Height="35" Margin="0,0,3,40" Click="PrevMedia_Click">
+            <Grid Style="{DynamicResource HoverEffect}" Background="#02000000">
+                <Path Data="{StaticResource IconArrowBordered}" Fill="White" RenderTransformOrigin="0.45,0.32">
+                    <Path.RenderTransform>
+                        <TransformGroup>
+                            <ScaleTransform ScaleX="1.5" ScaleY="1.5"/>
+                            <SkewTransform/>
+                            <RotateTransform Angle="-90"/>
+                            <TranslateTransform/>
+                        </TransformGroup>
+                    </Path.RenderTransform>
+                </Path>
+                <Grid.RenderTransform>
+                    <ScaleTransform/>
+                </Grid.RenderTransform>
+            </Grid>
+        </Button>
+        <Button Style="{StaticResource ButtonWrapper}" Cursor="Hand" VerticalAlignment="Center" HorizontalAlignment="Right" Width="35" Height="35" Margin="0,40,3,0" Click="NextMedia_Click">
+            <Grid Style="{DynamicResource HoverEffect}" Background="#02000000">
+                <Path Data="{StaticResource IconArrowBordered}" Fill="White" RenderTransformOrigin="0.32,0.45">
+                    <Path.RenderTransform>
+                        <TransformGroup>
+                            <ScaleTransform ScaleX="1.5" ScaleY="1.5"/>
+                            <SkewTransform/>
+                            <RotateTransform Angle="90"/>
+                            <TranslateTransform/>
+                        </TransformGroup>
+                    </Path.RenderTransform>
+                </Path>
+                <Grid.RenderTransform>
+                    <ScaleTransform/>
+                </Grid.RenderTransform>
+            </Grid>
+        </Button>
+        <Button Style="{StaticResource ButtonWrapper}" Cursor="Hand" VerticalAlignment="Bottom" HorizontalAlignment="Right" Width="35" Height="35" Margin="0,0,3,0" Click="MediaSliderFullscreen_Click">
+            <Grid Style="{DynamicResource HoverEffect}" RenderTransformOrigin="0.5,0.5" Background="#02000000">
+                <Path Data="{StaticResource IconGameViewMaximize}" Fill="White" RenderTransformOrigin="-0.05,-0.05">
+                    <Path.RenderTransform>
+                        <ScaleTransform ScaleX="2" ScaleY="2"/>
+                    </Path.RenderTransform>
+                </Path>
+                <Grid.RenderTransform>
+                    <ScaleTransform/>
+                </Grid.RenderTransform>
+            </Grid>
+        </Button>
+    </Grid>
+</UserControl>
